move record merge functionality into file, keep view functions
